pdf-bot/services/pdf_service/services/summarize_service.py
import pdfminer.high_level
import tempfile
import os
import logging
from pathlib import Path
from utils.response_utils import create_temp_response_file

logger = logging.getLogger(__name__)

class SummarizeService:
    """
    Service for PDF content summarization
    """
    
    def __init__(self):
        self.length_settings = {
            "short": {"sentences": 3, "max_words": 150},
            "medium": {"sentences": 5, "max_words": 300}, 
            "long": {"sentences": 10, "max_words": 600}
        }
    
    async def summarize(self, input_path: str, length: str = "medium", language: str = "en") -> str:
        """
        Summarize PDF content
        
        Args:
            input_path: Path to input PDF
            length: Summary length (short/medium/long)
            language: Output language
            
        Returns:
            Path to summary text file
        """
        try:
            logger.info(f"Summarizing PDF: {input_path}, length={length}, language={language}")
            
            # Extract text from PDF
            try:
                text = pdfminer.high_level.extract_text(input_path)
                if not text.strip():
                    raise ValueError("No text could be extracted from PDF")
            except Exception as e:
                logger.error(f"Text extraction failed: {e}")
                return self._create_placeholder_result(input_path, length, language)
            
            # Clean and prepare text
            cleaned_text = self._clean_text(text)
            
            # Generate summary (placeholder implementation)
            summary = self._generate_summary(cleaned_text, length, language)
            
            # Create output file
            return self._create_summary_output(summary, input_path, length)
            
        except Exception as e:
            logger.error(f"Summarization failed: {e}")
            return self._create_placeholder_result(input_path, length, language)
    
    def _clean_text(self, text: str) -> str:
        """Clean extracted text"""
        try:
            # Remove excessive whitespace
            cleaned = " ".join(text.split())
            
            # Remove common PDF artifacts
            cleaned = cleaned.replace("\x0c", " ")  # Form feed
            cleaned = cleaned.replace("\u2022", "•")  # Bullet points
            
            # Limit text length for processing (first 10000 characters)
            if len(cleaned) > 10000:
                cleaned = cleaned[:10000] + "..."
                
            return cleaned
            
        except Exception as e:
            logger.error(f"Text cleaning failed: {e}")
            return text
    
    def _generate_summary(self, text: str, length: str, language: str) -> str:
        """
        Generate summary (placeholder implementation)
        
        In a full implementation, this would use:
        - OpenAI GPT API
        - Hugging Face transformers
        - Local summarization models
        - Or other NLP services
        """
        try:
            settings = self.length_settings.get(length, self.length_settings["medium"])
            
            # Placeholder summarization logic
            # Split text into sentences
            sentences = [s.strip() for s in text.split('.') if s.strip()]
            
            # Take first N sentences as a simple summary
            summary_sentences = sentences[:settings["sentences"]]
            summary = ". ".join(summary_sentences)
            
            if not summary.endswith('.'):
                summary += "."
            
            # Add summary metadata
            summary_with_metadata = f"""DOCUMENT SUMMARY

Length: {length.capitalize()}
Target Language: {language.upper()}
Original Text Length: {len(text)} characters
Summary Length: {len(summary)} characters

SUMMARY:
{summary}

---
Note: This is a placeholder summarization. In production, this would use
advanced NLP models for proper content summarization.
"""
            
            return summary_with_metadata
            
        except Exception as e:
            logger.error(f"Summary generation failed: {e}")
            return f"Summary generation failed: {str(e)}"
    
    def _create_summary_output(self, summary: str, input_path: str, length: str) -> str:
        """Create summary output file"""
        try:
            # Add header information
            header = f"PDF Summary - {length.capitalize()}\n"
            header += f"Source: {Path(input_path).name}\n"
            header += f"Generated: {os.path.getctime(input_path)}\n"
            header += "=" * 50 + "\n\n"
            
            full_content = header + summary
            
            return create_temp_response_file(full_content, "txt")
            
        except Exception as e:
            logger.error(f"Failed to create summary output: {e}")
            raise
    
    def _create_placeholder_result(self, input_path: str, length: str, language: str) -> str:
        """Create placeholder result for testing"""
        try:
            placeholder_summary = f"""PDF SUMMARIZATION - PLACEHOLDER RESULT

Source File: {Path(input_path).name}
Summary Length: {length.capitalize()}
Target Language: {language.upper()}
Processing Status: Placeholder (actual AI summarization not available)

SUMMARY:
This is a placeholder summary for testing purposes. In a full implementation,
this would contain an AI-generated summary of the PDF content using advanced
natural language processing models.

The summary would be tailored to the requested length ({length}) and would
capture the key points, main arguments, and important details from the
original document.

Key features that would be included:
- Extractive and abstractive summarization techniques
- Content understanding and topic identification
- Language-specific processing for {language}
- Customizable summary length and focus areas

TECHNICAL IMPLEMENTATION NOTES:
- Text extraction using pdfminer.six
- NLP processing with transformers or OpenAI API
- Multi-language support for summaries
- Quality scoring and relevance ranking

---
Note: To enable full summarization functionality, integrate with:
- OpenAI GPT API for high-quality summaries
- Hugging Face transformers for local processing
- Google Cloud Natural Language API
- Or similar NLP services
"""
            
            return create_temp_response_file(placeholder_summary, "txt")
            
        except Exception as e:
            logger.error(f"Failed to create placeholder result: {e}")
            raise
