pdf-bot/services/pdf_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pikepdf==8.15.1
pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10
pdfminer.six==20231228
requests==2.31.0
python-docx==1.1.0
reportlab==4.0.7
httpx==0.25.2
pydantic==2.5.0
numpy==1.26.2
openpyxl==3.1.2
pydantic-settings==2.1.0
